fixed mapping errors, after login full window.location.reload get everything rendered into the right place, eliminated itemArray.map is not a function erro (i think)
